未保存変更管理を単一ソースに移動
